+++ src/NavBar/NavMenuItem.jsx
@@ -39,6 +39,7 @@ import { bool, number, string } from 'prop-types';
 import { useDispatch } from 'react-redux';
 import Button from 'react-bootstrap/Button';
 
+import classNames from '../utils/classNames';
 import { setCurrentPane } from '../App/appLayout';
 
 import './nav-menu-item.scss';
@@ -50,9 +51,11 @@ const NavMenuItem = ({ index, isFirst, isSelected, label }) => {
         <Button
             variant="link"
             active={false}
-            className={`core19-nav-menu-item ${isSelected ? 'selected' : ''} ${
-                isFirst ? 'first' : ''
-            }`}
+            className={classNames(
+                'core19-nav-menu-item',
+                isSelected && 'selected',
+                isFirst && 'first'
+            )}
             onClick={() => dispatch(setCurrentPane(index))}
             type="button"
         >