+++ src/Slider/Handle.jsx
@@ -35,13 +35,17 @@
  */
 
 import React, { useRef, useState } from 'react';
-import { func, number } from 'prop-types';
+import { bool, func, number } from 'prop-types';
+import classNames from '../utils/classNames';
+
 import rangeShape from './rangeShape';
 import {
     constrainedToPercentage,
     fromPercentage,
     toPercentage,
 } from './percentage';
+
+import './handle.scss';
 
 const useAutoupdatingRef = value => {
     const ref = useRef(value);
@@ -52,6 +56,7 @@ const useAutoupdatingRef = value => {
 const noop = () => {};
 const Handle = ({
     value,
+    disabled,
     range,
     onChange,
     onChangeComplete = noop,
@@ -101,11 +106,11 @@ const Handle = ({
 
     return (
         <div
-            className={`slider-handle ${currentlyDragged ? 'dragged' : ''}`}
+            className={classNames('handle', currentlyDragged && 'dragged')}
             style={{ left: `${percentage}%` }}
-            onMouseDown={grabHandle}
+            onMouseDown={disabled ? noop : grabHandle}
             role="slider"
-            tabIndex={0}
+            tabIndex={disabled ? -1 : 0}
             aria-valuemin={0}
             aria-valuemax={100}
             aria-valuenow={value}
@@ -114,6 +119,7 @@ const Handle = ({
 };
 Handle.propTypes = {
     value: number.isRequired,
+    disabled: bool.isRequired,
     range: rangeShape.isRequired,
     onChange: func.isRequired,
     onChangeComplete: func,